backend/ml_python/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.4
scikit-learn>=1.3.0
aiohttp>=3.9.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
transformers>=4.36.0
torch>=2.1.0
datasets>=2.14.0
better-profanity>=0.7.0
nltk>=3.8.0
sentence-transformers>=2.2.0